arquivos
Separei as classes já criadas até o momento e iniciei o estoque.

Algumas classes ainda estão juntas pois se relacionam de forma direta.

Vejam as mudanças e comentem se acham que deve separar mais ou juntar alguma classe com outra
