+++ produto.cpp
@@ -1,0 +1,30 @@
+#include "produto.hpp"
+#include <string>
+#include <iostream>
+#include <vector>
+#include <ctime>
+#include <cctype> 
+
+void Produto::alterar_valor(float valor_novo){
+   _valor = valor_novo;
+}
+
+void Produto::desconto(float porcentagem){
+   _valor = _valor - _valor*porcentagem;
+}
+
+std::string Produto::get_codigo(){
+    return _codigo_produto;
+}
+
+std::string Produto::get_nome(){
+    return _nome_produto;
+}
+
+float Produto::get_valor(){
+    return _valor;
+}
+
+std::string Produto::get_descricao(){
+    return _descricao;
+}
